undo changes on protos
