+++ app/controlplane/api/controlplane/v1/project.proto
@@ -51,6 +51,25 @@ message ProjectServiceListMembersResponse {
   OffsetPaginationResponse pagination = 2;
 }
 
+// ProjectMember represents an user or group who is a member of a project
+message ProjectMember {
+  // Subject can be either a user or a group
+  oneof subject {
+    // The user who is a member of the project
+    User user = 1;
+    // The group who is a member of the project
+    Group group = 2;
+  }
+  // The role of the user in the project
+  ProjectMemberRole role = 3;
+  // Timestamp when the project membership was created
+  google.protobuf.Timestamp created_at = 4;
+  // Timestamp when the project membership was last modified
+  google.protobuf.Timestamp updated_at = 5;
+  // The ID of latest project version this member is associated with
+  string latest_project_version_id = 6;
+}
+
 // ProjectServiceAddMemberRequest contains the information needed to add a user to a project
 message ProjectServiceAddMemberRequest {
   // IdentityReference is used to specify the project by either its ID or name
@@ -89,6 +108,16 @@ message ProjectMembershipReference {
     // The group to add to the project
     IdentityReference group_reference = 2;
   }
+}
+
+// ProjectMemberRole defines the roles a member can have in a project
+enum ProjectMemberRole {
+  // Default role for a project member
+  PROJECT_MEMBER_ROLE_UNSPECIFIED = 0;
+  // Admin role for a project member
+  PROJECT_MEMBER_ROLE_ADMIN = 1;
+  // Viewer role for a project member
+  PROJECT_MEMBER_ROLE_VIEWER = 2;
 }
 
 // ProjectServiceUpdateMemberRoleRequest contains the information needed to update a member's role in a project
