+++ app/controlplane/api/gen/frontend/controlplane/v1/project.ts
@@ -3,17 +3,55 @@ import { grpc } from "@improbable-eng/grpc-web";
 import { BrowserHeaders } from "browser-headers";
 import _m0 from "protobufjs/minimal";
 import { Timestamp } from "../../google/protobuf/timestamp";
+import { Group } from "./group";
 import { OffsetPaginationRequest, OffsetPaginationResponse } from "./pagination";
 import { User } from "./response_messages";
-import {
-  IdentityReference,
-  ProjectMember,
-  ProjectMemberRole,
-  projectMemberRoleFromJSON,
-  projectMemberRoleToJSON,
-} from "./shared_message";
+import { IdentityReference } from "./shared_message";
 
 export const protobufPackage = "controlplane.v1";
+
+/** ProjectMemberRole defines the roles a member can have in a project */
+export enum ProjectMemberRole {
+  /** PROJECT_MEMBER_ROLE_UNSPECIFIED - Default role for a project member */
+  PROJECT_MEMBER_ROLE_UNSPECIFIED = 0,
+  /** PROJECT_MEMBER_ROLE_ADMIN - Admin role for a project member */
+  PROJECT_MEMBER_ROLE_ADMIN = 1,
+  /** PROJECT_MEMBER_ROLE_VIEWER - Viewer role for a project member */
+  PROJECT_MEMBER_ROLE_VIEWER = 2,
+  UNRECOGNIZED = -1,
+}
+
+export function projectMemberRoleFromJSON(object: any): ProjectMemberRole {
+  switch (object) {
+    case 0:
+    case "PROJECT_MEMBER_ROLE_UNSPECIFIED":
+      return ProjectMemberRole.PROJECT_MEMBER_ROLE_UNSPECIFIED;
+    case 1:
+    case "PROJECT_MEMBER_ROLE_ADMIN":
+      return ProjectMemberRole.PROJECT_MEMBER_ROLE_ADMIN;
+    case 2:
+    case "PROJECT_MEMBER_ROLE_VIEWER":
+      return ProjectMemberRole.PROJECT_MEMBER_ROLE_VIEWER;
+    case -1:
+    case "UNRECOGNIZED":
+    default:
+      return ProjectMemberRole.UNRECOGNIZED;
+  }
+}
+
+export function projectMemberRoleToJSON(object: ProjectMemberRole): string {
+  switch (object) {
+    case ProjectMemberRole.PROJECT_MEMBER_ROLE_UNSPECIFIED:
+      return "PROJECT_MEMBER_ROLE_UNSPECIFIED";
+    case ProjectMemberRole.PROJECT_MEMBER_ROLE_ADMIN:
+      return "PROJECT_MEMBER_ROLE_ADMIN";
+    case ProjectMemberRole.PROJECT_MEMBER_ROLE_VIEWER:
+      return "PROJECT_MEMBER_ROLE_VIEWER";
+    case ProjectMemberRole.UNRECOGNIZED:
+    default:
+      return "UNRECOGNIZED";
+  }
+}
 
 /** ProjectServiceListMembersRequest contains the information needed to list members of a project */
 export interface ProjectServiceListMembersRequest {
@@ -29,6 +67,26 @@ export interface ProjectServiceListMembersResponse {
   members: ProjectMember[];
   /** Pagination information for the response */
   pagination?: OffsetPaginationResponse;
+}
+
+/** ProjectMember represents an user or group who is a member of a project */
+export interface ProjectMember {
+  /** The user who is a member of the project */
+  user?:
+    | User
+    | undefined;
+  /** The group who is a member of the project */
+  group?:
+    | Group
+    | undefined;
+  /** The role of the user in the project */
+  role: ProjectMemberRole;
+  /** Timestamp when the project membership was created */
+  createdAt?: Date;
+  /** Timestamp when the project membership was last modified */
+  updatedAt?: Date;
+  /** The ID of latest project version this member is associated with */
+  latestProjectVersionId: string;
 }
 
 /** ProjectServiceAddMemberRequest contains the information needed to add a user to a project */
@@ -272,6 +330,136 @@ export const ProjectServiceListMembersResponse = {
     message.pagination = (object.pagination !== undefined && object.pagination !== null)
       ? OffsetPaginationResponse.fromPartial(object.pagination)
       : undefined;
+    return message;
+  },
+};
+
+function createBaseProjectMember(): ProjectMember {
+  return {
+    user: undefined,
+    group: undefined,
+    role: 0,
+    createdAt: undefined,
+    updatedAt: undefined,
+    latestProjectVersionId: "",
+  };
+}
+
+export const ProjectMember = {
+  encode(message: ProjectMember, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
+    if (message.user !== undefined) {
+      User.encode(message.user, writer.uint32(10).fork()).ldelim();
+    }
+    if (message.group !== undefined) {
+      Group.encode(message.group, writer.uint32(18).fork()).ldelim();
+    }
+    if (message.role !== 0) {
+      writer.uint32(24).int32(message.role);
+    }
+    if (message.createdAt !== undefined) {
+      Timestamp.encode(toTimestamp(message.createdAt), writer.uint32(34).fork()).ldelim();
+    }
+    if (message.updatedAt !== undefined) {
+      Timestamp.encode(toTimestamp(message.updatedAt), writer.uint32(42).fork()).ldelim();
+    }
+    if (message.latestProjectVersionId !== "") {
+      writer.uint32(50).string(message.latestProjectVersionId);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): ProjectMember {
+    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseProjectMember();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          if (tag !== 10) {
+            break;
+          }
+
+          message.user = User.decode(reader, reader.uint32());
+          continue;
+        case 2:
+          if (tag !== 18) {
+            break;
+          }
+
+          message.group = Group.decode(reader, reader.uint32());
+          continue;
+        case 3:
+          if (tag !== 24) {
+            break;
+          }
+
+          message.role = reader.int32() as any;
+          continue;
+        case 4:
+          if (tag !== 34) {
+            break;
+          }
+
+          message.createdAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
+          continue;
+        case 5:
+          if (tag !== 42) {
+            break;
+          }
+
+          message.updatedAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
+          continue;
+        case 6:
+          if (tag !== 50) {
+            break;
+          }
+
+          message.latestProjectVersionId = reader.string();
+          continue;
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skipType(tag & 7);
+    }
+    return message;
+  },
+
+  fromJSON(object: any): ProjectMember {
+    return {
+      user: isSet(object.user) ? User.fromJSON(object.user) : undefined,
+      group: isSet(object.group) ? Group.fromJSON(object.group) : undefined,
+      role: isSet(object.role) ? projectMemberRoleFromJSON(object.role) : 0,
+      createdAt: isSet(object.createdAt) ? fromJsonTimestamp(object.createdAt) : undefined,
+      updatedAt: isSet(object.updatedAt) ? fromJsonTimestamp(object.updatedAt) : undefined,
+      latestProjectVersionId: isSet(object.latestProjectVersionId) ? String(object.latestProjectVersionId) : "",
+    };
+  },
+
+  toJSON(message: ProjectMember): unknown {
+    const obj: any = {};
+    message.user !== undefined && (obj.user = message.user ? User.toJSON(message.user) : undefined);
+    message.group !== undefined && (obj.group = message.group ? Group.toJSON(message.group) : undefined);
+    message.role !== undefined && (obj.role = projectMemberRoleToJSON(message.role));
+    message.createdAt !== undefined && (obj.createdAt = message.createdAt.toISOString());
+    message.updatedAt !== undefined && (obj.updatedAt = message.updatedAt.toISOString());
+    message.latestProjectVersionId !== undefined && (obj.latestProjectVersionId = message.latestProjectVersionId);
+    return obj;
+  },
+
+  create<I extends Exact<DeepPartial<ProjectMember>, I>>(base?: I): ProjectMember {
+    return ProjectMember.fromPartial(base ?? {});
+  },
+
+  fromPartial<I extends Exact<DeepPartial<ProjectMember>, I>>(object: I): ProjectMember {
+    const message = createBaseProjectMember();
+    message.user = (object.user !== undefined && object.user !== null) ? User.fromPartial(object.user) : undefined;
+    message.group = (object.group !== undefined && object.group !== null) ? Group.fromPartial(object.group) : undefined;
+    message.role = object.role ?? 0;
+    message.createdAt = object.createdAt ?? undefined;
+    message.updatedAt = object.updatedAt ?? undefined;
+    message.latestProjectVersionId = object.latestProjectVersionId ?? "";
     return message;
   },
 };
